requirements.txt
architect-functions==1.0.0
boto3==1.34.39
wasmtime=17.0.1
